+++ changelog.md
@@ -7,6 +7,7 @@
 * [fixed] Verification of media files
 * [changed] Most of the choice lists are sorted now
 * [added] Localities of Identity
+* [added] List of closest localities
 
 ## 02-2025